Fix bugs in node refutation functionality (https://github.com/runtimeverification/pyk/pull/877)
This PR fixes bugs in the implementation of node refutation, in advance
of implementing a `kontrol refute-node` command as part of
runtimeverification/evm-semantics#1752. The following changes are
introduced:
1. Fix the way node refutations are read from a JSON file:
1. When iterating over the `node_refutations` dictionary, the `.items()`
method was missing, meaning that only the keys were being iterated over,
rather than both keys and values.
2. In the same line, the node id needed to be converted back to an int
before being passed to `_resolve`, since the JSON file stores the key as
a string.
2. Prevent refuting a node that has successors (otherwise execution of
the refuted branch, which should have been paused, will continue from
the successor nodes).
3. Add `refuted` to the list of attributes for a node when displaying
the KCFG.

---------

Co-authored-by: Lucas MT <lucas.tabajara@runtimeverification.com>
Co-authored-by: devops <devops@runtimeverification.com>