Implement check for uses of deprecated symbols (#3991)
This PR implements a new `deprecated` attribute that can be applied to
productions; if that production is subsequently used in a rule the
compiler will emit a warning.

No code uses this feature yet, but I intend to use it to implement the
canonicalisation of `.` and `.K` (#3964, #3975) by marking `.` as
deprecated and doing some error-driven development.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>