Remove special-casing for non-unique symbols (#3975)
This PR is a partial fix for
https://github.com/runtimeverification/k/issues/3964; it makes the logic
for which productions are allowed to have duplicated symbol declarations
attached.

With a minimal amount of effort, all but one case of this special-casing
can be removed:
* One integration test still used the old `when` syntax for `requires`
clauses; as far as I can tell no actual code uses this syntax any more.
* Another test has two output stanzas permuted by some kind of ordering
nondeterminism.

The remaining case (empty K sequences) is harder to fix, as both
syntaxes are used in real K code. We have a few options here:
* Just permit this single special-casing to exist, and document that
both `.` and `.K` map to the KORE `dotk{}()`.
* Pick one of the two syntaxes as our preferred option, then add a
deprecation warning to the other and fix instances we find in real code.
Eventually remove the other one and make the check fully consistent with
no special cases.
* Implement a more general mechanism whereby two productions can produce
the same KORE.

Downstream testing to make sure nobody is actually using these old
cases:
* [x] C Semantics
* [x] KEVM

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>