Unpin LLVM backend Nixpkgs to allow an update to go through (#3992)
While we wait for the required changes to Nixpkgs etc. to go through
such that we can update the Haskell backend to a newer version, I think
it's worth allowing the two backends to diverge briefly such that the
LLVM backend update job can go through.

The divergence warning between the two versions of Nixpkgs successfully
triggers here.