Remove special case for array hook namespace (#4002)
For issue #3976 

The ARRAY module function symbols in `domains.md` have been hooked with
the ARRAY namespace, but no modern backends have those hooks implemented
so the frontend has been dropping the hook attributes with specific
checks, falling back to the `ARRAY-IN-K` semantics for them.

This change removes the ARRAY hook attributes in `domains.md` and the
special cases for them in the frontend.