Update regression tests to use `.K` rather than `.` (#4006)
This is an initial change to minimise any fallout from making `.` a
deprecated symbol; it's a bulk change of uses of the unadorned `.` in
the K regression test suite to instead use `.K`.

Once this is merged, we can start to think about doing a similar change
for downstream semantics, then finally deprecating and removing the
symbol.

Part of: https://github.com/runtimeverification/k/issues/3964