Add attribute-based overloading (#3985)
This PR is the next step towards the changes suggested in
https://github.com/runtimeverification/k/issues/3966; having added a new
syntax for the `symbol(_)` attribute in #3978, we can now address the
second use of `klabel(_)` (marking overloaded sets of symbols).

The changes in this PR are as follows:
* Add a new attribute `overload(_)` that behaves identically to the
current semantics of `klabel(_)` with respect to the construction of the
overload partial ordering, but that does not participate in the _naming_
of symbols.
* Add checks for the new attribute:
* Disallow `klabel(_)` and `overload(_)` on the same production;
overload sets must be converted fully to use the new mechanism.
  * Warn when an overload set of only one production is generated.
* Document the semantics of overloading in the user manual based on our
conversations in #3966.
* Add a new integration test that checks a simple example of overloading
works as expected to disambiguate parses.

Once this PR is merged, we can begin the long process of deprecating
`klabel(_)` and `symbol`!

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>