Support Java 17 in `kore` submodule (#4032)
In #3736, we updated to Java 17 by configuring the Java version for
`maven-compiler-plugin`.

However - after banging my head against a wall for the past hour
wondering why none of my code is parsing 🙃 - I realized that the `kore`
and `matching` submodules are actually built by `scala-maven-plugin`
rather than `maven-compiler-plugin`, and their Java version must be set
independently.

This PR sets the version for `kore` to Java 17, and
runtimeverification/llvm-backend#993 will do the same for `matching`.