Check for uses of zero-argument `symbol` without `klabel(_)` (#4044)
Fixes: #4043
Related: #4035, #3979  

This PR adds a compiler warning when `symbol` is used without an
argument and without a corresponding `klabel(_)` attribute.

I fixed one case in `domains.md` where this was an issue; for
consistency I left that production using `klabel(_)` rather than
migrating it to `symbol(_)`. We will have to go back and fix these cases
anyway in a later PR.