Consolidate GenerateCoverage logic (#4040)
#3973 

This defers adding K-IO to the main module from parsing to when the
coverage instrumentation gets generated. It also makes
GenerateCoverage's methods static, and moves the check for the coverage
option outside of it, where generating coverage just turns into a no-op
if the flag isn't there.

There still needs to be a check in DefinitionParsing on the coverage
flag to keep the K-IO module around at the step where unused modules get
trimmed from the definition.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>