Optimize KCFG memory usage (https://github.com/runtimeverification/pyk/pull/872)
This optimization works by making sure that a single object is used for
all groups of equal subterms in a kcfg's nodes.

This is implemented by keeping a map from (optimized) subterms to IDs,
and another map from subterm IDs to actual subterms. Whenever a new node
is added to the kcfg, this PR will do a bottom-up traversal of its
subterms, replacing them with their cached version, if that exists, or
adding them to the cache otherwise.

Memory usage
---------------

For the add liquidity proof, when running with Kasmer, the kcfg object
uses 50-60GB of memory. With this PR, it uses only 1 GB.

Educated guess: In general, I expect that the memory usage for proofs
whose configurations do not change much between consecutive nodes
(usually that would be configurations of medium or high complexity), the
memory usage is higher when the kcfg holds 1-2 nodes, and lower
afterwards. For proofs whose configurations change significantly between
nodes (usually that would mean simple configurations, e.g. with only the
`<k>` cell), this PR may increase the memory usage.

Runtime
--------

* Running a simple proof (i.e. most of the time is spent in the Booster
backend): 6 min 59 sec with the PR, 6 min 56 sec before PR
* Loading a small kcfg (29M on disk) from json: 5.5 sec with the PR, 4.9
sec before PR
* Loading a large kcfg (6.1G on disk) from json: 1606 sec with the PR,
1425 sec before PR

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>
Co-authored-by: Lisandra <lisandrasilva18@hotmail.com>
Co-authored-by: Noah Watson <107630091+nwatson22@users.noreply.github.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>