Fix KCFG.replace_node functionality (#920)
Fixes: https://github.com/runtimeverification/pyk/issues/686

Currently, `KCFG.replace_node` is broken because `KCFG.Successor` stores
directly a `KCFG.Node` for the sources and targets. So when we replace a
node in the store currently, we are only replacing the node itself, but
not iterating through the successors and replacing occurances of the
same node there.

This PR uses a pretty brute-force way to do that (to be able to make
progress on storing Node attributes on nodes directly instead of as
separate lists), and later we can make a better way of storing this
information: https://github.com/runtimeverification/pyk/issues/919.

- A test is added, that is failing on master but passing on this PR,
that retrieving a node that is part of an edge happens correctly after
erplacement.
- Helpers `KCFG.Succesor.{replace_source,replace_target}` are added for
producing new copies of the successors with th eappropriate node on them
updated.
- The routine `KCFG.replace_node` is updated to use these replacer
methods and update the actual edge dictionaries.

---------

Co-authored-by: devops <devops@runtimeverification.com>