Remove KPrint object from KCFGExplore (#928)
Currently, the `KCFGExplore` object holds both a `KPrint` and
`CTermSymbolic`. But it really only needs `KPrint` for the
`pretty_print` functionality, which only really relies on the
`PrettyPrinter` class in the `kast/pretty` module. So this PR reduces
that dependency a bit.

- Places where `KCFGExplore` needed a `KDefinition` now reuse the one
held by `CTermSymbolic`.
- `KCFGExplore._pretty_printer` is added for building a `PrettyPrinter`
directly. `KCFGExplore.pretty_print` is added to expose the pretty
printing functionality of that `PrettyPrinter`.
- Tests and other files are updated.

---------

Co-authored-by: devops <devops@runtimeverification.com>