Add `to_claim` method to `RefutationProof` (https://github.com/runtimeverification/pyk/pull/933)
This PR adds a `to_claim` method to `RefutationProof` which returns a
claim representing the refutation. This claim is of the form
```
claim ['claim-id']: LAST_CONSTRAINT => false
  requires PRE_SPLIT_CONSTRAINTS [label(claim-id)]
```
where `LAST_CONSTRAINT` is the branch condition immediately preceding
the refuted node and `PRE_SPLIT_CONSTRAINTS` is the set of path
constraints that depend (directly or indirectly) on the variables of
`LAST_CONSTRAINT`.

It also makes the following refactors to support this implementation:
* Refactor `apply_existential_substitutions` to take configuration and
constraints separately, rather than as a conjoined term, and return them
separately as well. This allows avoiding `split_config_and_constraints`,
which requires the config to be wrapped in a cell.
* Refactor `build_claim` and `build_rule` to take configuration and
constraints separately, rather than packed in a `CTerm`, and creates
separate `cterm_build_claim` and `cterm_build_rule` wrappers for the
`CTerm` version.

---------

Co-authored-by: devops <devops@runtimeverification.com>