Factor out KCFGStore class (https://github.com/runtimeverification/pyk/pull/945)
Part of: https://github.com/runtimeverification/pyk/issues/919

This PR is the first in a series to begin having a more explicit
`KCFGStore` object for actually storing/maintaining the data associated
with a KCFG rather than directly have `KCFG` do it. This PR:

- Deduplicates some code in `KCFG.write_cfg_data` by reusing
`KCFG.to_dict`.
- Factors out `KCFGStore` class from `KCFG`, and gives `KCFG` a
`KCFGStore` object for now to maintain backwards compatibility with
existing interface (including wiring up the previous functions in `KCFG`
to the new ones in `KCFGStore`).
- Adjusts the tests of the KCFG file storage.

This PR aims to maintain backwards compatibility, while giving a
factored out class to work with.

---------

Co-authored-by: devops <devops@runtimeverification.com>