Restore pretty-printing for exhaustiveness
Previously, this logic lived in the LLVM backend's matching compiler.
When we eliminated the backwards dependency edge between the backend and
the frontend, this meant we couldn't use the frontend pretty-printing
logic to make the error message in the pattern matcher.

This commit reinstates equivalent pretty-printing logic in the frontend.