Emit better error message for locally-duplicated syntactic lists (#4084)
Previously, we would only inspect _imported_ sentences when looking for
duplicated user lists. This means that if you duplicate the production
locally, you hit an assertion instead of the existing check.

To fix this, the check needs to ensure that it sees multiple `userList`
sentences _with different_ source / location pairs_ for a given sort.

Fixes: #4079