Add RPC based prover to KProve interface (https://github.com/runtimeverification/pyk/pull/948)
This PR aims to expose a very basic RPC prover in the `KProve` class. In
particular:

- The command `pyk prove ...` is renamed to `pyk prove-legacy ...` (and
single test adjusted). This functionality is not really used anywhere.
- The functionality `legacy_explore` from KEVM is upstreamd to
`cterm_symbolic` helper in `cterm.symbolic`. It has the same option to
start or not an RPC server, and also implements the context-manager
functionality.
- The functionality to do proving with RPC server is added to `KProve`.
We have:
- `KProve.prove_claim_rpc`: starts up a `kore-rpc` server using
`cterm_symbolic`, and runs the proof as an `APRProof`. Nothing fancy is
done with interpreting claim dependencies, circularities, or
parallelism.
- `KProve.prove_rpc`: mimics the `KProve.prove` interface where a spec
file with a potential list of claims is passed in, collects the claims
and discharges them with `KProve.prove_claim_rpc`.
- A lot of defunct functionality is removed from `KProve`:
- Early prototype of RPC based prover `KProve.prove_cterm` is removed,
and associated tests.
- Various arguments to `KProve.prove` which were not being used
anywhere, or only used in specific ways, are eliminated to simplify the
function (for eventual removal of that function, as it's the
`prove-legacy` hookup).
- The function `KProve.get_claim_modules` no longer uses
`KProve.prove(...)` functionality, instead directly invokes
`_kprove(...)` helper to further allow `KProve.prove(...)` to be
simplified.
- A test-suite is added for `KProve.prove_rpc` which re-uses all the
tests from `proof/test_imp.py` which are passing using the new interface
directly, to make sure the functionality doesn't break.

---------

Co-authored-by: devops <devops@runtimeverification.com>