Set `failure_info` if `fail_fast` is on (#962)
Unblocks https://github.com/runtimeverification/kontrol/pull/429.

This PR sets `failure_info` for a failed proof even if the `fail_fast`
option is on and the proof terminates early.

It also adds a `fail_fast` parameter and a corresponding test claim to
`test_failure_info` (but not
[`test_failure_info_recomputed_on_proof_reinit`](https://github.com/runtimeverification/pyk/blob/a442ceaebeef0872de26749d1513fc37d9675206/src/tests/integration/proof/test_imp.py#L1070)
for the reason mentioned in the
[comment](https://github.com/runtimeverification/pyk/pull/962#issuecomment-1988023070)).

---------

Co-authored-by: devops <devops@runtimeverification.com>