Port `POSet` to Java (#4103)
Part of #4030.

Port the `POSet` class from Scala to Java. 

This is mostly a straightforward conversion from Scala functional idioms
(`map`, `filter`, `reduce`, etc.) to the corresponding Java Stream
methods. To port `lazy val`s, we also implement a `Lazy<T>` wrapper
which caches the result of a `Supplier<T>`.

The only remaining reference to Scala here is a single constructor which
takes a `scala.collection.Set<Tuple2<T, T>>` and internally converts it
to a `java.util.Set<Pair<T, T>>`. This can be removed once we more
pervasively switch away from Scala collection types everywhere in the
codebase.

---------

Co-authored-by: Bruce Collie <brucecollie82@gmail.com>