Add class `POSet` (https://github.com/runtimeverification/pyk/pull/991)
Adds a class `POSet` that takes a binary relation, computes it's
transitive closure and stores the image of the resulting relation. The
sort ordering is re-implemented to use `POSet` for both KAST and KORE.

Note that technically, that the name `Proset` would be more fitting for
the current implementation as neither antisymmetry, nor asymmetry is
enforced. (Reflexivity isn't either, but that's easy to work around.)

---------

Co-authored-by: devops <devops@runtimeverification.com>