Reapply "Miscellaneous Scala clean up" (#4078) (#4108)
This reverts commit ec70555ac575bdba60ae59c389de3bebbdc47509, which was
itself a revert of #4055 because of breakages in our Nix builds because
of issues around the `scala-kore` dependency. This PR reinstates the
original change now that `scala-kore` has been properly released and
versioned, and also makes a fix (see comment below) for an issue with
Nix / Maven integration that was producing spurious hash invalidations.

---------

Co-authored-by: Scott Guest <scott.guest@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>