Implement check for non-unique overload sets (#4117)
This PR adds a compiler warning when multiple distinct overload sets use
the same key. Doing so is not ambiguous from the compiler's perspective
(arity and sorting are sufficient to distinguish the different
components of the partial orders), but it may be confusing to a user.

We couldn't easily implement this check previously because two
productions could legitimately have the same `klabel(_)`.

Fixes https://github.com/runtimeverification/k/issues/4034