Fix userList edge case from #4117 (#4121)
In #4117, we added a warning when two structurally-different productions
share an overload attribute. This produced false-positive warnings on
user list productions, as these productions are desugared internally
into two separate cons / nil productions that will share an overload
attribute based on their original production.

This PR fixes the edge case by checking if precisely 2 productions, all
with the `userList` attribute, share an overload key. In this case, we
don't emit a warning. The test case from the original PR has been
updated with a negative example as a regression test.