Add total attribute to projection functions for named nonterminals which are total (#4118)
As part of
https://github.com/runtimeverification/evm-semantics/pull/2349 we are
trying to change how we implement the gas schedule in KEVM in order to
improve performance. In order to do this, we introduce a new sort,
ScheduleTuple, which has a single constructor with many arguments. We
rely on projection functions for named nonterminals in order to extract
those arguments. However, this confuses the booster because it is unable
to determine that these functions are total. In general, they are not
total, however, for the case where the sort they are defined over has a
single constructor, they are total. Thus, we introduce logic into
GenerateSortProjections which tags these productions as total in the
cases where they are in fact total.

This has been tested and shown to fix the performance regression in the
booster introduced by the above change. It should be easy to see why
this change is sound since it applies the `total` attribute only in the
case where the production has a single constructor, and the rule matches
unconditionally on any instance of that constructor.