Don't generate `inj{S, S}` when symbol overloads have the same sort (#4136)
It is valid for two overloaded symbols to have the same sort; a
motivating case for doing so is to select a _more specific function
implementation_ based on more specific arguments, rather than to make
the sort of a constructor more specific. We do this a few times in the C
semantics, for example.

Previously, such overloads would generate spurious injections `inj{S,
S}` in the "overloaded symbol" axioms. Because we don't actually rewrite
based on these axioms (but rather just read off the relevant
attributes), this behaviour went unrecognised for a long time.

For consistency, this PR amends ModuleToKore such that it will not
generate the injection when the sort of the two productions is the same.
We're in the process of porting this code to Python, and so I have not
made any effort to clean up the code here; doing so is really outside
the scope of this change. I have added a test that minimally exercises
this part of the KORE generator, and verifies that the overload set
induced works correctly to disambiguate a parse.

Fixes: https://github.com/runtimeverification/k/issues/4114