Separate pre- and post-compilation checks (#4138)
When working on #4045, I noticed that a pair of productions like:
```k
syntax KItem ::= disambiguate(KItem, KItem) [overload(disambiguate)]
               | disambiguate(List,  Set)   [overload(disambiguate)]
```
were both being incorrectly flagged as singleton overloads, despite
meeting the criteria for being an overload set.

Some investigation reveals that this is due to the overload lattice
checks being applied _before_ the compiler inserts subsorting
productions into `KItem` for every sort; we therefore don't know that
`List <: KItem` when constructing the overload lattice.

The solution is to move the overload checks later in the compilation
process, after the compilation pipeline has been run. This PR makes that
change, along with some associated documentation and renaming for
clarity. It also includes a regression test with the example above; I
have verified that the test breaks before this PR.