Remove manip.get_cell, add CTerm.free_vars (https://github.com/runtimeverification/pyk/pull/1040)
Inspired while reviewing:
https://github.com/runtimeverification/pyk/pull/1009/

This makes two changes which hopefully tighten the `CTerm` abstraction
slightly by reducing the places where `CTerm.kast` is used:

- Places that were doing `get_cell(CTerm.kast, CELL_NAME)` now say
`CTerm.cell(CELL_NAME)`. There are no more uses of `get_cell` here or in
downstream repos, so it's removed.
- A new `@cached_property CTerm.free_vars` is added, which is used to
replace the pattern `free_vars(CTerm.kast)`.

This means that the remaining uses of `CTerm.kast` are doing something
weird with the resulting kast, or printing it, or converting it to Kore.

---------

Co-authored-by: devops <devops@runtimeverification.com>