Add check for non-symbol productions with `overload(_)` attribute (#4143)
For full context, see the original issue (#4142). This PR implements a
simple check for overloaded non-symbol productions, along with a minimal
test case.

Also included in the PR is an example minimised from the C Semantics
that helps to explain _why_ we would want to overload on subsort
productions (and therefore, why the fix for this issue is a check rather
than, say, forbidding the feature).

In brief, the new integration test minimises part of the C expression
grammar, in which we have the following sort hierarchy:
```
    Exp
     |
    Val
   /   \
RVal   LVal
```
We also wish to construct a corresponding hierarchy of left-associative,
comma-separated syntactic lists over this hierarchy (with the exception
being that there is no need for a list of lvalues):
```
    Exps
     |
    Vals
   /
RVals
```

Typically, the solution here would be to implement the list sorts as K
syntactic lists. For example:
```
syntax Exps ::= NeList{Exp, ","}
```

This is, of course, just syntactic sugar, and we can write down an
equivalent set of productions:
```
syntax Exps ::= Exp
              | Exps "," Exp
```

Note here that the singleton production is a _subsorting_ production
that would, by default, not have an associated `klabel`, and therefore
would not be generated as a KORE symbol.

However, a collision with a separate part of the C grammar does indeed
prevent us from using the `NeList` sugar for this, as we require
finer-grained control over the parsing priorities. The issue is with the
comma operator:
```
syntax Exp ::= Exp "," Exp [left, group(comma)]
```

We require the property that each element in the overloaded list-sort
hierarchy has a higher priority than the comma operator, and so we need
to be able to attach groups to both the singleton and comma productions.

When combined with the necessary overloading attributes on the singleton
list subsorting productions, this situation triggers the bug fixed by
this PR.

Fixes https://github.com/runtimeverification/k/issues/4142