Exporting `get_requires` utils function (https://github.com/runtimeverification/pyk/pull/1056)
This PR introduces a new file, `utils.py`, which should contain
auxiliary functions that need to be converted from and to llvm but don't
set any field on an object.

In this context we're using a function from llvm backend's
[pattern_matching.cpp](https://github.com/runtimeverification/llvm-backend/blob/e2f583a05b9bf362deabd58f8acc1ce534aaf4c4/lib/ast/pattern_matching.cpp#L174)
to avoid reimplement it in Python. This function is exported as Pyhton
binding in this
[PR](https://github.com/runtimeverification/llvm-backend/pull/1019), and
therefore, the current PR can only be merged after we merge the LLVM
Backend one.

---------

Co-authored-by: devops <devops@runtimeverification.com>