Don't emit no-junk symbol for macro symbols (#4156)
Previously the no junk axiom for symbols included symbols that had the
`macro` attribute. This is incorrect since these symbols are not
constructors. When we fix this, the attached test begins to pass,
whereas previously it would be unable to encode the `size` symbol to
SMT, and thus unable to discharge the constraint, causing it to branch
incorrectly.