Align Pyk version with K
K and Pyk are now one homogeneous release, so we need to align the
version of the Pyk library with the top-level K version, and bump it on
pushes to develop as we do for the debian package and changelog.

I have tested the sed expression used in the version bumper locally and
it produces the correct behaviour.

This commit also removes Pyk's own packaging files; these are no longer
needed as we're opting in to using the top-level K version of everything
here.