Remove proof field from Prover (#4250)
Porting https://github.com/runtimeverification/pyk/pull/960 to K repo.

`APRProver.step_proof` currently assumes the proof being passed in the
`APRProofStep` it takes as argument is the same as `self.proof`, i.e.
when it calls `self.nonzero_depth`, `self._check_subsume`, and
`self._checked_for_bounded` assumes the proof being passed in is always
the same for a single `APRProver` instance.

This is a problem in itself, but changing this should also move us
closer an implementation that can run `step_proof` on different nodes in
parallel, by reducing its dependency on external data. A next step could
be initializing a new KCFGExplore every `step_proof` instead of using
`self.kcfg_explore` to access the `KoreClient`.

- Removes `proof` field from `Prover` and `APRProver`.
- Makes methods that read/write `self.prover` take a prover argument 
- Proof initialization is performed separately in `init_proof` at the
beginning of `advance_proof` rather than in the `Prover` constructor.
- Moves `_checked_for_terminal` and `_checked_for_bounded` into
`APRProof`, as they refer to nodes of a specific proof.