Add `parse-outer` command to pyk (#4270)
This starts implementing a command with pyk that handles the outer
parsing of a K definition into a KAST JSON that can be passed to inner
parsing in the java frontend. So far it can handle requires clauses of
other files, do markdown selection, and handle extra include paths
passed in from the command line.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>