Use release tags for LLVM backend (#4304)
This PR moves the LLVM backend over to using a release tag workflow
rather than a submodule one; doing so makes it easier to implement
planned changes to our CI workflows.

I have tested the changes here by running them on this branch and
rebasing out the changes made.

This PR should not be merged until
https://github.com/runtimeverification/devops/pull/209 is approved (but
not merged), so that the devops PR can happen immediately on
`llvm-backend_release` appearing in this repo.