Set RTLD_LOCAL via context manager (#4306)
See
https://github.com/runtimeverification/k/issues/4302#issuecomment-2089079706
for a full breakdown of the reason for this change, and the original
issue that caused it to manifest for me.

In short, there is a difference between how Linux and macOS handle
loading dynamic libraries that look like KLLVM runtime bindings modules.
The Linux behaviour (setting `RTLD_LOCAL` when loading Python native
extension modules via `dlopen`) is what we want, and so the solution is
to ensure that behaviour happens unconditionally across all platforms.

I've verified that this fixes my issue locally; we can't test this
properly without doing a bigger refactoring of our Nix code, so I'd like
to merge this and return to the test suite improvements subsequently.

Fixes #4302