Increase heap size in frontend compiler (#4320)
The Kontrol team are experiencing OOMs on larger projects
(https://github.com/runtimeverification/k-private-issues/issues/9); the
solution for now is to be more generous with the heap size that K's
allowed to use.

We are currently working on a refactoring of the compilation pipeline;
once that's done we can think about being more parsimonious with our
memory usage.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>