Support NULL return from `requires` binding of `get_requires()` (#4326)
The llvm function `get_requires()` can return either the Pattern in the
requires clause or nullptr. However, our binding only accepted the first
case, and this PR fixes that behavior by testing if the result of
`axiom.requires` returned a pattern. If not, we return None instead of
trying to convert None to a LLVM Pattern.

A unit test was added to ensure this behavior was documented and tested.