Fix flaky un-alias test (#4331)
In https://github.com/runtimeverification/k/pull/4300, I added new tests
that reproduced an issue from Kontrol. Unfortunately, one of those test
cases overlapped with the existing test case for this code. As a result,
the test became flaky (maybe 1 in 10 runs on my machine) would fail
because the `#range` and `rangeHundred` aliases overlapped for the `0 <=
X < 100` case.

The fix is just to remove the overlapping rule ported in from Kontrol; I
only added the 4 different rules because it was a direct copy-paste. We
are still testing the regressive behaviour fixed in #4300 equally well
because there are still multiple `#range` alias symbols that should be
considered by `indexed_rewrite`.

I have verified as far as possible that the test no longer flakes on my
machine.