Source and Location attributes for outer parsed modules (#4313)
The Java frontend is going to be intolerant of missing source and
location attributes in various places after outer parsing, so we need to
make sure any KAST that we pass to it when skipping its outer parser has
them.

This implements those attributes for modules, taking a step towards
getting the initial end to end compilation with pyk's outer parser.