Fix `import_kllvm` in Pyk and modifying imports in ProofTrace (#4339)
This PR aims to fix the `import_kllvm` function to look for `_kllvm` in
`sys.modules` before importing it from a file. This unblocks the Pi2
team from using the ProofTrace wrappers for the Python Bindings.

This PR also modifies the ProofTrace module to use the optimized version
of `pyk.kllvm.load` and its test so as not to reimport it.