Minimizing branching constraints (#4299)
In response to [issue
#3836](https://github.com/runtimeverification/haskell-backend/issues/3836)
in the Haskell backend, this PR minimizes the set of branching
constraints produced by the backend in the following way:
1. It isolates an intersection of all of the branching constraints
provided - if this intersection is non-empty, then the branching should
not depend on it (as otherwise there should be branches where its
negation holds);
2. It checks whether this intersection is entailed by the current path
condition, and:
a. if so, removes the intersection from all of the branching
constraints;
    b. if not, does not make any changes. 

This solution is sound, but not complete, since we check the
intersection as a whole rather than each constraint individually because
each check is a call to the SMT solver. However, since this check is
done for only for branching results and only for cases in which such an
intersection exists, which happens rarely, perhaps we could improve by
check individually.

I have tested this with Kontrol and I can confirm that the spurious
constraints described in [issue
#3836](https://github.com/runtimeverification/haskell-backend/issues/3836)
no longer appear.