Add textual KAST parser to `pyk` (#4336)
Known differences to the JavaCC implementation:
* Tokens `.::K` and `.::KList` are not supported.
* Token `#klabel` is not supported.
* When lexing `.K` and `.KList`, the lexer reads until special
character. This allows a simple LA(1) implementation. On the other hand,
e.g. `.KA` will raise an error.
* This double comma won't parse:

https://github.com/runtimeverification/k/blob/7be107f58bb97c8f50989b60247c0180f80d88c8/k-frontend/src/main/javacc/KastParser.jj#L182
* When unquoting strings (`"Int"`) and klabels (`` `<k>` ``), a more
straightforward decoding is used that simply replaces `\X` by `X`. This
is in line with how these tokens are lexed.

Potential improvements to the format:
* Support `?X`.
* Support `X:Int`.
* Support klabels with sort arguments.
* Write `foo` or `foo()` instead of `foo(.KList)`.
* Write something like `Int("1")` (like an application) or `"1":Int`
(like a variable) instead of `#token("1", "Int")` (though there might be
a reason `KToken` is represented that way.)