Fix incorrect mvnHash value (#4367)
I'm not sure why this has broken; best guess is that an upstream
dependency has done a force-push to a Maven repository in a way that
causes a JAR / POM to be overwritten.

The same new hash value was appearing on every failing PR run, and on
`develop` when I built K locally. This suggests pretty strongly to me
that this isn't something we've broken, but rather an upstream issue.