kompilex command for pyk (#4342)
Implements the command `pyk kompilex` which uses pyk's outer parser to
parse in a K Definition and then hands it off to the java frontend to
handle the rest of the compilation.

Some bootstrapping is required because the conversions from the outer
parsed structures to something compatible with the java frontend aren't
fully implemented. For the integration tests, a `prelude-modules.json`
file is included in the test data folder which has the K prelude already
outer parsed.

This will only handle modules that are empty or import other modules.
The goal is to expand this capability until the full K language is
supported, and then the bootstrapping will no longer be necessary.

---------

Co-authored-by: Bruce Collie <brucecollie82@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>