CI: Don't tell pytest to try to kill the integration tests on failure (#4371)
There's a bug with `pytest-xdist` where using multiple runners doesn't
properly terminate the test run when there's a failure and `--maxfail=1`
has been passed. Instead, the runner that ran the failed test exits and
the remaining runners continue. What this results in on CI is the
remaining runners are not able to finish the integration tests before
the timeout for the job, killing pytest before it can finish and report
the reason for the failed test.

This change removes the `--maxfail` option when running on CI, so when
there's a test failure, all of the runners continue, so they should be
able to finish before the timeout and pytest can properly finish and
report the failure reason(s).

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>