Run homebrew release on macOS 14 / arm64 (#4373)
We are working on dropping support for intel macOS across K-related
projects; one of the last load-bearing things we relied on that platform
for was running Homebrew builds and tests. As of this year, Github offer
hosted macOS 14 / arm64 runners that solve the problem we ran into
previously where our own hosted runners were not ephemeral enough to
properly run these builds.

This PR moves the relevant parts of our release workflows over to the
new hosted runners; this code isn't directly tested in the PR flow so I
opened a dummy PR that runs the relevant parts of the workflow. You can
see the tests passing here:
https://github.com/runtimeverification/k/pull/4372