Adding new proof trace tests from Pi2 (#4421)
This PR adds a set of new tests that previously lived in the Pi2
repository but no longer need to be there as the company now uses the
LLVM Hints Parser instead of its own parser. Therefore, we decided to
move the tests from the Pi2 repo to PyK so it can not only increase our
test suit but also be as close as possible to the source code.

We decided to keep the whole Hints input instead of using PyK to build
it, as we realized that not only were the traces generated using the
corresponding PyK function different but also smaller than we expected,
modifying the test behavior. One of the differences we spot is that
`map_pattern` adds an unnecessary `\left-assoc` in hints production that
modifies the final trace.

To facilitate the review of this PR, each new test corresponds to a
commit!