Drop remaining uses of Intel macOS in CI (#4430)
We've been working through the CI setups for the other K components and
removing support for Intel macOS; this PR finishes the process by
removing support from K itself.

There are a couple of places where we can simplify conditional checks in
workflow steps, and one place where it makes sense to run a job on
`macos-14` (which will be arm64).