Align Pyk's nixpkgs version to the shared K version (#4439)
The Pyk Nix flake was using a floating version of Nixpkgs; this PR pins
it to share the same version as the rest of the K toolchain.

Doing this will help us simplify lots of downstream code that consumes K
/ Pyk via Nix; several projects end up in a situation where they need to
instantiate multiple Nixpkgs to get Pyk _and_ K working.

This is a preparatory change that I want to make before bringing Pyk
into the main K flake.