Adding new PyK wrappers to LLVM Python Bindings (#4426)
This PR exposes the new infrastructure created by the LLVm Backend to
parse the Hints file.
We introduce the wrappers to make the streaming parser possible to use
as a list of events and access it through an integrator. Or not loading
all the events at the same time but accessing them one per time through
the `next()` function.

Depends on:
https://github.com/runtimeverification/llvm-backend/pull/1090

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>