Reorganise CI to use pre-packaged LLVM backend release (#4308)
This PR refactors the way that we build Docker images with K installed
to run later CI stages (i.e. the Pyk integration tests). Rather than
building the entire K distribution from scratch, we instead download and
install a pre-packaged release of the LLVM backend. A subsequent (much
smaller) change will do the same for the Haskell backend, now that the
supporting infrastructure is in place.

I don't expect that CI times will actually decrease meaningfully with
this PR; the LLVM backend is pretty quick to build. Doing the same for
the haskell backend will be a much bigger impact, and should bring CI
times down to ~20 minutes.

The existing CI job that produces a "whole of K" distribution is still
run, and we still release that package on new K versions.